python/requirements.txt
discord.py
psycopg2
oauth2client
requests
BeautifulSoup4
lxml
psycopg2

